I'm so close, everything is copying, do not know why it isn't properly formatting

+++ src/main.py
@@ -8,10 +8,11 @@ from pagegenerator import generate_page, generate_pages_recursive
 def main():
 	public_dir = "/home/mici/gitHub/MicahsProjects/staticWebsite/public"
 	template_dir = "/home/mici/gitHub/MicahsProjects/staticWebsite/template.html"
-	from_dir = "/home/mici/gitHub/MicahsProjects/staticWebsite/content/index.html"
+	from_dir = "/home/mici/gitHub/MicahsProjects/staticWebsite/content/"
 	static_dir = "/home/mici/gitHub/MicahsProjects/staticWebsite/static"
 	print("Hello world! Executing main.py!")
-	shutil.rmtree(public_dir)
+	if os.path.exists(public_dir):
+		shutil.rmtree(public_dir)
 	directory_copy(static_dir, public_dir)
 	generate_pages_recursive(from_dir,template_dir,public_dir)
 	subprocess.run(["python3", "-m", "http.server", "8888"], cwd=public_dir)
